requirements.txt
absl-py==0.10.0
aplus==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.1
astropy==4.3.post1
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
blake3==0.2.0
bleach==3.2.1
bqplot==0.12.30
CacheControl==0.12.6
cachetools==4.1.1
cachy==0.3.0
certifi==2020.6.20
cffi==1.14.3
cfgv==3.3.1
chardet==3.0.4
cleo==0.8.1
click==7.1.2
clikit==0.6.2
cloudpickle==1.6.0
colorama==0.4.3
crashtest==0.3.1
cycler==0.10.0
dask==2021.7.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.2.10
django-google-maps==0.12.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
frozendict==2.0.3
fsspec==2021.7.0
future==0.18.2
gast==0.3.3
gevent==20.9.0
google-auth==1.21.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
html5lib==1.1
identify==2.2.13
idna==2.10
importlib-metadata==1.7.0
ipydatawidgets==4.2.0
ipykernel==5.3.4
ipyleaflet==0.14.0
ipympl==0.7.0
ipython==7.18.1
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipyvue==1.5.0
ipyvuetify==1.8.0
ipywebrtc==0.6.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keyring==21.8.0
kiwisolver==1.2.0
lab==6.1
llvmlite==0.36.0
locket==0.2.1
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
msgpack==1.0.2
mysql-connector-python==8.0.23
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
nodeenv==1.6.0
notebook==6.1.4
numba==0.53.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
partd==1.2.0
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.7.1
poetry==1.1.8
poetry-core==1.0.4
pre-commit==2.14.0
progressbar2==3.53.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyerfa==2.0.0
pyforest==1.1.0
Pygments==2.7.1
PyJWT==2.3.0
pylev==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-utils==2.5.6
pythreejs==2.3.0
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.1
Send2Trash==1.5.0
shellingham==1.4.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
tensorflow-hub==0.9.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tomlkit==0.7.2
toolz==0.11.1
tornado==6.0.4
traitlets==5.0.4
traittypes==0.2.1
txt2tags==3.7
typing_extensions==4.0.1
urllib3==1.25.10
vaex==4.3.0
vaex-astro==0.8.2
vaex-core==4.3.0.post1
vaex-hdf5==0.8.0
vaex-jupyter==0.6.0
vaex-ml==0.12.0
vaex-server==0.5.0
vaex-viz==0.5.0
virtualenv==20.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.19.0
zipp==3.2.0
zope.event==4.5.0
zope.interface==5.1.0
